requirements.txt
ess-streaming-data-types >= 0.7.1
kafka-python
pre-commit
pytest
